move CoxxFingerprint to featurize.utils
